Refactor edit of exception days
The patch contains no functional changes, but is a necessary precondition for
extentions of update_rept() (in ui-day.c) with further recurrence rules.

The reason is that recurrence parameters must be treated as a whole: if an edit
session is cancelled at any point, no value should change, and all parameters
should remain as they were. Hence, the new values must only be set after all of
them have been determined. This was not the case for the list of exception days,
but as long as it was treated last, it did not matter.