Improve data load error reporting
The last part of loading appointments and events is performed by four "scan"
functions called from io_load_app(). Failure in this part of data load does not
use io_load_error().

The four "scan" functions are changed to return an error message on failure and
NULL otherwise (the previous return value was not used).