User interface for recurrence rules
The function update_rept() is extended with editing of the three recurrence rule
lists for BYMONTH, BYMONTHDAY and BYDAY.

The integers of the bymonth and bymonthday lists are edited directly as
integers, while those of the bywday list are mapped to localized weekday names
(as they appear in the calendar panel) with an optional integer prefix (in
RFC5545 style: 1MO, -2SA).

The RFC5545 (icalendar) requirement that the start day must be the first
occurence and must match the recurrence rule, is met by testing that an
occurence indeed appears on the start day, in these circumstances:
	- when a recurrent item is loaded from file
	- when the recurrence rule of an item is edited interactively
	- when a recurrent appointment gets a new start time
	- when a recurrent appointment is moved

Copy and paste of a recurrent item will only retain the basic recurrence
properties of type, frequency, until and exception days.