Treat recurrent events and appointments alike
The day vector (or "day_items" vector) is used to load the appointments panel
with "day_items" (captions, appointments, events). A multiday appointment has
several "day_items" in the day vector. This may also be the case for a recurring
item.

The day_item structure has a member "start", which for day_items for recurring
appointments is set to the occurrence found by recur_item_find_occurrence(), and
is used to tell days of a multiday appointment apart.

For day_items for recurring events it was set to "day" of the recur_event
structure (the same value for all occurrences) and not used. The value is now
the occurrence found by recur_event_find_occurrence().

The patch is backwards compatible and future-proof.

Signed-off-by: Lars Henriksen <LarsHenriksen@get2net.dk>