Update llist.c
When ical/ics files are imported, omit duplicate entries. Useful for stand-alone calendars on Android and desktops so each can synchronize by exporting/importing their entire calendars to each other repeatedly. The patch prints the details on each entry that is omitted. This behavior could be suppressed with a command-line option.