Remove SIGCHLD signal handler
The purpose is to make child_wait() reliable. The handler is meant for the
notify main thread, but signal handlers are shared by all threads. In the
calcurse main thread and the periodic save thread (hooks) the handler and
child_wait() will compete (if the signal handler kicks in first, the waitpid()
call in child_wait() will fail with errno ECHILD).

All child processes in the main thread, the notify thread, the periodic save
thread and the notify demon are explicitly waited for.

Signed-off-by: Lars Henriksen <LarsHenriksen@get2net.dk>