calcurse-caldav: when pushing events, always take href as returned by server
Apparently, while some caldav servers will return href values as is in
their response (as seen in issues #337 and #356), some might return them
url-escaped (which if I am not mistaken was the reason for e943b06).

Assuming either behaviour when pushing local objects will lead to
discrepancy with events dictionary retrieved from the server and thus
bugs, hence we always want to use whatever form of href the server
returns.