Formats status message strings longer than cols
Previously if msg1 of a status_mesg was longer than the number of
columns, it would be overwritten by msg2.

Now if msg1 is longer than the number of columns, msg2 is offset by 5
from the end of msg1, and a vertical bar character '|' is placed to show
the separation between lines. If the 5 character separation would place
msg2 offscreen, the separation is not applied.

Signed-off-by: Morgan Seltzer <MorganSeltzer000@gmail.com>