sha1: Fix checksum calculation on big-endian systems
Instead of using a custom macro which must be set manually to determine
host byte order (WORDS_BIGENDIAN) this commit uses the __BYTE_ORDER__
CPP macro which is commonly available on many systems [1]. If the macro
is not defined, little endian byte order is silently assumed which is
what the previous version of this code did anyhow. Alternatively, it
would also be possible to emit a #warning if __BYTE_ORDER__ is not set
or rewrite the code in a way that it does not depend on the native host
byte order in the first place [2].

This fixes the calcurse test suite on s390x.

Fixes #397

[1]: https://gcc.gnu.org/onlinedocs/cpp/Common-Predefined-Macros.html
[2]: https://commandcenter.blogspot.com/2012/04/byte-order-fallacy.html