remove unnecessary set of sidebar width
When the function is called, `col` is zero, which leads to setting the
sidebar width (in percent) to zero. As the variable holding the
percentage is `static unsigned sbarwidth_perc` it is initialized to zero
anyway and we set the percentage later to the value loaded from the
configuration, too.

Signed-off-by: Max <max@mxzero.net>
Signed-off-by: Lukas Fleischer <lfleischer@calcurse.org>