calcurse-caldav: use flock to prevent concurrent instances
The current locking implementation has two issues:

1. There is a race condition between lock file creation and checking for
   lock file existence.

2. A process holding the lock can die without removing the file.

This commit fixes both problems with a new locking implementation. To
obtain the lock, a process obtains an exclusive flock on the lock file,
creating the file if it doesn't exist. The kernel takes care of removing
the flock once the locked descriptor is closed (i.e., when the process
exits).

Signed-off-by: Lukas Fleischer <lfleischer@calcurse.org>